Specify when pushstate routing is set up
This updates the logic that determines when pushstate routing is used.
Previously it was specifically avoided if using the file: protocol
because this is the protocol that Cordova and NW.js use. Now that NW.js
uses chrome-extension: as the protocol, the currently logic doesn't
work.

This makes the logic more flexible by having a whitelist of protocols
that are valid to set up pushstate routing. Those protocols are:

* http
* https
* (empty string) - This one is for done-ssr, because can-simple-dom
* doesn't have a protocol.

Fixes #29