Fix an issue with the tests passing in IE11
The failing test was expecting that an anchor element’s `pathname` would start with `/`, but IE11 has a bug where it doesn’t include the leading `/` if the element isn’t part of the document: https://stackoverflow.com/questions/956233/javascript-pathname-ie-quirk

This fixes the issue by adding the element to the document in the test. I think this was a better solution than changing the core `can-route-pushstate` code to detect the situation, because under normal use I would only expect links to be clicked if they’re part of a document.